+++ examples/resize.rs
@@ -1,13 +1,19 @@
 extern crate doryen_rs;
 
-use doryen_rs::{App, AppOptions, DoryenApi, Engine, TextAlign};
+use doryen_rs::{App, AppOptions, DoryenApi, Engine, TextAlign, UpdateEvent};
 
 struct MyRoguelike {
     width: u32,
     height: u32,
+    mouse_pos: (f32, f32),
 }
 
 impl Engine for MyRoguelike {
+    fn update(&mut self, api: &mut dyn DoryenApi) -> Option<UpdateEvent> {
+        let input = api.input();
+        self.mouse_pos = input.mouse_pos();
+        None
+    }
     fn render(&mut self, api: &mut dyn DoryenApi) {
         let con = api.con();
         con.rectangle(
@@ -16,7 +22,7 @@ impl Engine for MyRoguelike {
             self.width,
             self.height,
             Some((128, 128, 128, 255)),
-            None,
+            Some((0, 0, 0, 255)),
             Some(' ' as u16),
         );
         con.area(
@@ -36,6 +42,11 @@ impl Engine for MyRoguelike {
             None,
             None,
         );
+        con.back(
+            self.mouse_pos.0 as i32,
+            self.mouse_pos.1 as i32,
+            (255, 255, 255, 255),
+        );
     }
     fn resize(&mut self, api: &mut dyn DoryenApi) {
         self.width = api.get_screen_size().0 / 8;
@@ -49,6 +60,7 @@ impl MyRoguelike {
         Self {
             width: 80,
             height: 50,
+            mouse_pos: (0.0, 0.0),
         }
     }
 }