+++ src/console.rs
@@ -112,6 +112,9 @@ impl Console {
     pub fn get_height(&self) -> u32 {
         self.height
     }
+    pub fn get_size(&self) -> (u32, u32) {
+        (self.width, self.height)
+    }
     pub fn get_pot_width(&self) -> u32 {
         self.pot_width
     }